apiapp/requirement.txt
Django==2.1.3
django-filter==2.0.0
djangorestframework==3.9.0
numpy==1.16.1
pandas==0.24.1
pi==0.1.2
python-dateutil==2.7.5
pytz==2018.4
six==1.12.0
sqlparse==0.3.0